scripting/javascript/bindings/js/debugger/tests/mochitest/test_inspector-changeattrs.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Components.utils.import("resource://gre/modules/devtools/Loader.jsm");

const Promise = devtools.require("sdk/core/promise");
const inspector = devtools.require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gInspectee = null;
var gClient = null;
var gWalker = null;
var checkActorIDs = [];

function assertOwnership() {
  assertOwnershipTrees(gWalker);
}

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = devtools.require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addTest(function testChangeAttrs() {
  let attrNode = gInspectee.querySelector("#a");
  let attrFront;
  promiseDone(gWalker.querySelector(gWalker.rootNode, "#a").then(front => {
    attrFront = front;
    dump("attrFront is: " + attrFront + "\n");
    // Add a few attributes.
    let list = attrFront.startModifyingAttributes();
    list.setAttribute("data-newattr", "newvalue");
    list.setAttribute("data-newattr2", "newvalue");
    return list.apply();
  }).then(() => {
    // We're only going to test that the change hit the document.
    // There are other tests that make sure changes are propagated
    // to the client.
    is(attrNode.getAttribute("data-newattr"), "newvalue", "Node should have the first new attribute");
    is(attrNode.getAttribute("data-newattr2"), "newvalue", "Node should have the second new attribute.");
  }).then(() => {
    // Change an attribute.
    let list = attrFront.startModifyingAttributes();
    list.setAttribute("data-newattr", "changedvalue");
    return list.apply();
  }).then(() => {
    is(attrNode.getAttribute("data-newattr"), "changedvalue", "Node should have the changed first value.");
    is(attrNode.getAttribute("data-newattr2"), "newvalue", "Second value should remain unchanged.");
  }).then(() => {
    let list = attrFront.startModifyingAttributes();
    list.removeAttribute("data-newattr2");
    return list.apply();
  }).then(() => {
    is(attrNode.getAttribute("data-newattr"), "changedvalue", "Node should have the changed first value.");
    ok(!attrNode.hasAttribute("data-newattr2"), "Second value should be removed.");
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gWalker;
  delete gInspectee;
  delete gClient;
  runNextTest();
});


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
